Fix unit tests failing due to supabase-client import chain
Extract pure functions (buildReviewsMap, getIndexOfNextUnreviewedCard,
getIndexOfNextAgainCard) and types (ReviewsMap, ReviewStages) into
review-utils.ts which has no supabase dependency. Move useNextValid
from store.ts to hooks.ts. This breaks the store.ts → hooks.ts →
supabase-client.ts import chain that caused unit tests to fail with
"supabaseUrl is required" when env vars aren't set.
