+++ src/routes/_user/learn/$lang.review.go.tsx
@@ -4,13 +4,12 @@ import { createFileRoute, Navigate } from '@tanstack/react-router'
 import { ChevronLeft, ChevronRight } from 'lucide-react'
 import {
 	useCardIndex,
-	useNextValid,
 	useReviewActions,
 	useReviewDayString,
 	useReviewLang,
 	useReviewStage,
 } from '@/features/review/store'
-import { useReviewDay } from '@/features/review/hooks'
+import { useNextValid, useReviewDay } from '@/features/review/hooks'
 import { Loader } from '@/components/ui/loader'
 import { WhenComplete } from '@/components/review/when-review-complete-screen'
 import { ReviewSingleCard } from '@/components/review/review-single-card'
